refactor: move `getCurrentCurrency`` from `ui/selectors/selectors.js`` to `ui/ducks/metamask/metamask.js``
