+++ ui/components/app/assets/token-cell/token-cell.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 import { BigNumber } from 'bignumber.js';
+import { getCurrentCurrency } from '../../../../ducks/metamask/metamask';
 import {
-  getCurrentCurrency,
   getTokenList,
   selectERC20TokensByChain,
   getNativeCurrencyForChain,
