+++ ui/components/app/wallet-overview/aggregated-percentage-overview.tsx
@@ -3,8 +3,8 @@ import { useSelector } from 'react-redux';
 
 import { toChecksumAddress } from 'ethereumjs-util';
 import { getNativeTokenAddress } from '@metamask/assets-controllers';
+import { getCurrentCurrency } from '../../../ducks/metamask/metamask';
 import {
-  getCurrentCurrency,
   getSelectedAccount,
   getShouldHideZeroBalanceTokens,
   getTokensMarketData,
