+++ ui/components/multichain/token-list-item/price/percentage-and-amount-change/percentage-and-amount-change.test.tsx
@@ -5,13 +5,13 @@ import { zeroAddress } from 'ethereumjs-util';
 import { MarketDataDetails } from '@metamask/assets-controllers';
 import { getIntlLocale } from '../../../../../ducks/locale/locale';
 import {
-  getCurrentCurrency,
   getSelectedAccountCachedBalance,
   getTokensMarketData,
 } from '../../../../../selectors';
 import { getCurrentChainId } from '../../../../../../shared/modules/selectors/networks';
 import {
   getConversionRate,
+  getCurrentCurrency,
   getNativeCurrency,
 } from '../../../../../ducks/metamask/metamask';
 import { PercentageAndAmountChange } from './percentage-and-amount-change';
@@ -25,7 +25,6 @@ jest.mock('../../../../../ducks/locale/locale', () => ({
 }));
 
 jest.mock('../../../../../selectors', () => ({
-  getCurrentCurrency: jest.fn(),
   getSelectedAccountCachedBalance: jest.fn(),
   getTokensMarketData: jest.fn(),
 }));
@@ -35,6 +34,7 @@ jest.mock('../../../../../../shared/modules/selectors/networks', () => ({
 }));
 
 jest.mock('../../../../../ducks/metamask/metamask', () => ({
+  getCurrentCurrency: jest.fn(),
   getConversionRate: jest.fn(),
   getNativeCurrency: jest.fn(),
 }));
