+++ ui/ducks/metamask/metamask.js
@@ -604,3 +604,7 @@ export function doesUserHaveALedgerAccount(state) {
     return kr.type === KeyringType.ledger;
   });
 }
+
+export function getCurrentCurrency(state) {
+  return state.metamask.currentCurrency;
+}
