+++ ui/hooks/useAccountTotalCrossChainFiatBalance.ts
@@ -2,6 +2,9 @@ import { shallowEqual, useSelector } from 'react-redux';
 import { toChecksumAddress } from 'ethereumjs-util';
 import {
   getCurrentCurrency,
+  getCurrencyRates,
+} from '../ducks/metamask/metamask';
+import {
   getCrossChainTokenExchangeRates,
   getCrossChainMetaMaskCachedBalances,
 } from '../selectors';
@@ -9,7 +12,6 @@ import {
   getValueFromWeiHex,
   sumDecimals,
 } from '../../shared/modules/conversion.utils';
-import { getCurrencyRates } from '../ducks/metamask/metamask';
 import { getTokenFiatAmount } from '../helpers/utils/token-util';
 import { TokenWithBalance } from '../components/app/assets/asset-list/asset-list';
 import { getNetworkConfigurationsByChainId } from '../../shared/modules/selectors/networks';
