+++ ui/hooks/useAccountTotalFiatBalance.js
@@ -3,7 +3,6 @@ import { toChecksumAddress } from 'ethereumjs-util';
 import { getCurrentChainId } from '../../shared/modules/selectors/networks';
 import {
   getAllTokens,
-  getCurrentCurrency,
   getMetaMaskCachedBalances,
   getTokenExchangeRates,
   getConfirmationExchangeRates,
@@ -18,6 +17,7 @@ import {
 import {
   getConversionRate,
   getNativeCurrency,
+  getCurrentCurrency,
 } from '../ducks/metamask/metamask';
 import { formatCurrency } from '../helpers/utils/confirm-tx.util';
 import { getTokenFiatAmount } from '../helpers/utils/token-util';
