+++ ui/hooks/useFiatFormatter.ts
@@ -1,6 +1,6 @@
 import { useSelector } from 'react-redux';
 import { getIntlLocale } from '../ducks/locale/locale';
-import { getCurrentCurrency } from '../selectors';
+import { getCurrentCurrency } from '../ducks/metamask/metamask';
 import { shortenString } from '../helpers/utils/util';
 
 /**
