+++ ui/hooks/useTokensToSearch.js
@@ -6,13 +6,14 @@ import { formatIconUrlWithProxy } from '@metamask/assets-controllers';
 import { getTokenFiatAmount } from '../helpers/utils/token-util';
 import {
   getTokenExchangeRates,
-  getCurrentCurrency,
   getSwapsDefaultToken,
   getTokenList,
 } from '../selectors';
 import { getCurrentChainId } from '../../shared/modules/selectors/networks';
-import { getConversionRate } from '../ducks/metamask/metamask';
-
+import {
+  getConversionRate,
+  getCurrentCurrency,
+} from '../ducks/metamask/metamask';
 import { getSwapsTokens } from '../ducks/swaps/swaps';
 import { isSwapsDefaultTokenSymbol } from '../../shared/modules/swaps.utils';
 import { toChecksumHexAddress } from '../../shared/modules/hexstring-utils';
